Convert matrices over polynomials to matrices over the complex numbers
This is a bit of a downer, since having matrices over polynomials
actually worked quite nicely, and allowed us to not worry about floating
point arithmetic until very late. Unfortunately, when we did have to
worry about that, we would also have to worry a great deal, since even
for simple evaluations, like q = -1, the coefficients would be so large
for relatively long words that floating point arithmetic would break in
the case of integers. We fix this by just evaluating immediately
instead, at the expense of floating point errors potentially
accumulating.